% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{flode_theme}
\alias{flode_theme}
\title{Construct a FlodeTheme configuration object}
\usage{
flode_theme(
  base_size = 11,
  base_family = "",
  palette = "flood",
  show_grid = TRUE
)
}
\arguments{
\item{base_size}{Base font size in points. Default \code{11}.}

\item{base_family}{Base font family string. Default \code{""}.}

\item{palette}{One of \code{"flood"} or \code{"minimal"}. Default \code{"flood"}.}

\item{show_grid}{Include major gridlines. Default \code{TRUE}.}
}
\value{
A \code{FlodeTheme} S7 object.
}
\description{
Construct a FlodeTheme configuration object
}
\examples{
ft <- flode_theme()
ft@palette_name

}
