% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{print.FlodeTheme}
\alias{print.FlodeTheme}
\title{S3 print method for FlodeTheme}
\usage{
\method{print}{FlodeTheme}(x, ...)
}
\arguments{
\item{x}{A \code{FlodeTheme} object.}

\item{...}{Unused.}
}
\description{
S3 print method for FlodeTheme
}
