% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_qc}
\alias{scale_colour_qc}
\title{Colour scale for QC flag integers (1–6)}
\usage{
scale_colour_qc(na.value = "#767676", name = "QC flag", ...)
}
\arguments{
\item{na.value}{Colour for NA values. Default grey.}

\item{name}{Legend title. Default \code{"QC flag"}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Maps QC flag codes defined in the Bronze/Silver schema to the EA flood
information palette. Intended for use with a \code{colour} aesthetic mapped to an
integer QC flag column.
}
\examples{
library(ggplot2)
library(data.table)
dt <- data.table(
  t    = seq.POSIXt(Sys.time(), by = "hour", length.out = 10),
  val  = runif(10, 0, 5),
  flag = sample(1:4, 10, replace = TRUE)
)
ggplot(dt, aes(t, val, colour = factor(flag))) +
  geom_point() +
  scale_colour_qc()

}
