% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_condition}
\alias{scale_fill_condition}
\title{Fill scale for model condition ratings}
\usage{
scale_fill_condition(na.value = "#767676", name = "Condition", ...)
}
\arguments{
\item{na.value}{Colour for NA values.}

\item{name}{Legend title. Default \code{"Condition"}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Fill scale for model condition ratings
}
