% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_qc}
\alias{scale_fill_qc}
\title{Fill scale for QC flag integers (1–6)}
\usage{
scale_fill_qc(na.value = "#767676", name = "QC flag", ...)
}
\arguments{
\item{na.value}{Colour for NA values. Default grey.}

\item{name}{Legend title. Default \code{"QC flag"}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Fill scale for QC flag integers (1–6)
}
