% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_tier}
\alias{scale_fill_tier}
\title{Fill scale for asset tiers}
\usage{
scale_fill_tier(name = "Tier", ...)
}
\arguments{
\item{name}{Legend title. Default \code{"Tier"}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Fill scale for asset tiers
}
